test: Move files
